+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added cardsummary to fdd fault in myems-web
 - added bind energy storage power station to space in myems-admin
 - added energy plan reports in myems-api
+- added advanced settings menu to myems-admin
 ### Changed
 - changed map marker in myems-web
 - updated popup style of mapbox in myems-web